refactor: bottom navigation bar
