+++ lib/screens/home_screen.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:trips_app/routes.dart' as routes;
+
+import 'package:trips_app/widgets/common/bottom_navigation_bar.dart';
+// import 'package:trips_app/routes.dart' as routes;
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -9,26 +11,6 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  int _currentTabIndex = 0;
-
-  void _onTabBottomNavigationBarItem(int index) {
-    _changeScreen(index);
-    setState(() {
-      _currentTabIndex = index;
-    });
-  }
-
-  void _changeScreen(int index) {
-    switch (index) {
-      case 0:
-        Navigator.pushNamed(context, routes.homeScreen);
-        break;
-      case 1:
-        Navigator.pushNamed(context, routes.placeScreen);
-        break;
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -37,24 +19,7 @@ class _HomeScreenState extends State<HomeScreen> {
           'Home screen',
         ),
       ),
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentTabIndex,
-        items: [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.home),
-            label: "Home",
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.place),
-            label: "Lugares",
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.account_circle),
-            label: "Cuenta",
-          )
-        ],
-        onTap: _onTabBottomNavigationBarItem,
-      ),
+      bottomNavigationBar: BottomNavigationBarWidget(),
     );
   }
 }
